add tests for price write
