+++ core/services/ocr2/plugins/ccip/ccipcommit/ocr2_test.go
@@ -3,10 +3,8 @@ package ccipcommit
 import (
 	"context"
 	"encoding/json"
-	"fmt"
 	"math/big"
 	"math/rand"
-	"reflect"
 	"slices"
 	"sort"
 	"testing"
@@ -1094,210 +1092,6 @@ func TestCommitReportingPlugin_calculatePriceUpdates(t *testing.T) {
 	}
 }
 
-func TestCommitReportingPlugin_generatePriceUpdates(t *testing.T) {
-	val1e18 := func(val int64) *big.Int { return new(big.Int).Mul(big.NewInt(1e18), big.NewInt(val)) }
-
-	const nTokens = 10
-	tokens := make([]cciptypes.Address, nTokens)
-	for i := range tokens {
-		tokens[i] = cciptypes.Address(utils.RandomAddress().String())
-	}
-	sort.Slice(tokens, func(i, j int) bool { return tokens[i] < tokens[j] })
-
-	testCases := []struct {
-		name                 string
-		tokenDecimals        map[cciptypes.Address]uint8
-		sourceNativeToken    cciptypes.Address
-		priceGetterRespData  map[cciptypes.Address]*big.Int
-		priceGetterRespErr   error
-		feeEstimatorRespFee  *big.Int
-		feeEstimatorRespErr  error
-		maxGasPrice          uint64
-		expSourceGasPriceUSD *big.Int
-		expTokenPricesUSD    map[cciptypes.Address]*big.Int
-		expErr               bool
-	}{
-		{
-			name: "base",
-			tokenDecimals: map[cciptypes.Address]uint8{
-				tokens[0]: 18,
-				tokens[1]: 18,
-			},
-			sourceNativeToken: tokens[0],
-			priceGetterRespData: map[cciptypes.Address]*big.Int{
-				tokens[0]: val1e18(100),
-				tokens[1]: val1e18(200),
-				tokens[2]: val1e18(300), // price getter returned a price for this token even though we didn't request it (should be skipped)
-			},
-			priceGetterRespErr:   nil,
-			feeEstimatorRespFee:  big.NewInt(10),
-			feeEstimatorRespErr:  nil,
-			maxGasPrice:          1e18,
-			expSourceGasPriceUSD: big.NewInt(1000),
-			expTokenPricesUSD: map[cciptypes.Address]*big.Int{
-				tokens[0]: val1e18(100),
-				tokens[1]: val1e18(200),
-			},
-			expErr: false,
-		},
-		{
-			name: "price getter returned an error",
-			tokenDecimals: map[cciptypes.Address]uint8{
-				tokens[0]: 18,
-				tokens[1]: 18,
-			},
-			sourceNativeToken:   tokens[0],
-			priceGetterRespData: nil,
-			priceGetterRespErr:  fmt.Errorf("some random network error"),
-			expErr:              true,
-		},
-		{
-			name: "price getter skipped a requested price",
-			tokenDecimals: map[cciptypes.Address]uint8{
-				tokens[0]: 18,
-				tokens[1]: 18,
-			},
-			sourceNativeToken: tokens[0],
-			priceGetterRespData: map[cciptypes.Address]*big.Int{
-				tokens[0]: val1e18(100),
-			},
-			priceGetterRespErr: nil,
-			expErr:             true,
-		},
-		{
-			name: "price getter skipped source native price",
-			tokenDecimals: map[cciptypes.Address]uint8{
-				tokens[0]: 18,
-				tokens[1]: 18,
-			},
-			sourceNativeToken: tokens[2],
-			priceGetterRespData: map[cciptypes.Address]*big.Int{
-				tokens[0]: val1e18(100),
-				tokens[1]: val1e18(200),
-			},
-			priceGetterRespErr: nil,
-			expErr:             true,
-		},
-		{
-			name: "base",
-			tokenDecimals: map[cciptypes.Address]uint8{
-				tokens[0]: 18,
-				tokens[1]: 18,
-			},
-			sourceNativeToken: tokens[0],
-			priceGetterRespData: map[cciptypes.Address]*big.Int{
-				tokens[0]: val1e18(100),
-				tokens[1]: val1e18(200),
-				tokens[2]: val1e18(300), // price getter returned a price for this token even though we didn't request it
-			},
-			priceGetterRespErr:   nil,
-			feeEstimatorRespFee:  big.NewInt(10),
-			feeEstimatorRespErr:  nil,
-			maxGasPrice:          1e18,
-			expSourceGasPriceUSD: big.NewInt(1000),
-			expTokenPricesUSD: map[cciptypes.Address]*big.Int{
-				tokens[0]: val1e18(100),
-				tokens[1]: val1e18(200),
-			},
-			expErr: false,
-		},
-		{
-			name: "dynamic fee cap overrides legacy",
-			tokenDecimals: map[cciptypes.Address]uint8{
-				tokens[0]: 18,
-				tokens[1]: 18,
-			},
-			sourceNativeToken: tokens[0],
-			priceGetterRespData: map[cciptypes.Address]*big.Int{
-				tokens[0]: val1e18(100),
-				tokens[1]: val1e18(200),
-				tokens[2]: val1e18(300), // price getter returned a price for this token even though we didn't request it (should be skipped)
-			},
-			priceGetterRespErr:   nil,
-			feeEstimatorRespFee:  big.NewInt(20),
-			feeEstimatorRespErr:  nil,
-			maxGasPrice:          1e18,
-			expSourceGasPriceUSD: big.NewInt(2000),
-			expTokenPricesUSD: map[cciptypes.Address]*big.Int{
-				tokens[0]: val1e18(100),
-				tokens[1]: val1e18(200),
-			},
-			expErr: false,
-		},
-		{
-			name: "nil gas price",
-			tokenDecimals: map[cciptypes.Address]uint8{
-				tokens[0]: 18,
-				tokens[1]: 18,
-			},
-			sourceNativeToken: tokens[0],
-			priceGetterRespData: map[cciptypes.Address]*big.Int{
-				tokens[0]: val1e18(100),
-				tokens[1]: val1e18(200),
-				tokens[2]: val1e18(300), // price getter returned a price for this token even though we didn't request it (should be skipped)
-			},
-			feeEstimatorRespFee: nil,
-			maxGasPrice:         1e18,
-			expErr:              true,
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			priceGetter := pricegetter.NewMockPriceGetter(t)
-			defer priceGetter.AssertExpectations(t)
-
-			gasPriceEstimator := prices.NewMockGasPriceEstimatorCommit(t)
-			defer gasPriceEstimator.AssertExpectations(t)
-
-			var destTokens []cciptypes.Address
-			for tk := range tc.tokenDecimals {
-				destTokens = append(destTokens, tk)
-			}
-			sort.Slice(destTokens, func(i, j int) bool {
-				return destTokens[i] < destTokens[j]
-			})
-			var destDecimals []uint8
-			for _, token := range destTokens {
-				destDecimals = append(destDecimals, tc.tokenDecimals[token])
-			}
-
-			queryTokens := ccipcommon.FlattenUniqueSlice([]cciptypes.Address{tc.sourceNativeToken}, destTokens)
-
-			if len(queryTokens) > 0 {
-				priceGetter.On("TokenPricesUSD", mock.Anything, queryTokens).Return(tc.priceGetterRespData, tc.priceGetterRespErr)
-			}
-
-			if tc.maxGasPrice > 0 {
-				gasPriceEstimator.On("GetGasPrice", mock.Anything).Return(tc.feeEstimatorRespFee, tc.feeEstimatorRespErr)
-				if tc.feeEstimatorRespFee != nil {
-					pUSD := ccipcalc.CalculateUsdPerUnitGas(tc.feeEstimatorRespFee, tc.expTokenPricesUSD[tc.sourceNativeToken])
-					gasPriceEstimator.On("DenoteInUSD", mock.Anything, mock.Anything).Return(pUSD, nil)
-				}
-			}
-
-			p := &CommitReportingPlugin{
-				sourceNative:      tc.sourceNativeToken,
-				priceGetter:       priceGetter,
-				gasPriceEstimator: gasPriceEstimator,
-			}
-
-			destPriceReg := ccipdatamocks.NewPriceRegistryReader(t)
-			destPriceReg.On("GetTokensDecimals", mock.Anything, destTokens).Return(destDecimals, nil).Maybe()
-			p.destPriceRegistryReader = destPriceReg
-
-			sourceGasPriceUSD, tokenPricesUSD, err := p.generatePriceUpdates(context.Background(), logger.TestLogger(t), destTokens)
-			if tc.expErr {
-				assert.Error(t, err)
-				return
-			}
-			assert.NoError(t, err)
-			assert.True(t, tc.expSourceGasPriceUSD.Cmp(sourceGasPriceUSD) == 0)
-			assert.True(t, reflect.DeepEqual(tc.expTokenPricesUSD, tokenPricesUSD))
-		})
-	}
-}
-
 func TestCommitReportingPlugin_isStaleReport(t *testing.T) {
 	ctx := context.Background()
 	lggr := logger.TestLogger(t)
@@ -1615,46 +1409,6 @@ func Test_calculateIntervalConsensus(t *testing.T) {
 			}
 			assert.Equal(t, tt.wantMin, got.Min)
 			assert.Equal(t, tt.wantMax, got.Max)
-		})
-	}
-}
-
-func Test_calculateUsdPer1e18TokenAmount(t *testing.T) {
-	tests := []struct {
-		name       string
-		price      *big.Int
-		decimal    uint8
-		wantResult *big.Int
-	}{
-		{
-			name:       "18-decimal token, $6.5 per token",
-			price:      big.NewInt(65e17),
-			decimal:    18,
-			wantResult: big.NewInt(65e17),
-		},
-		{
-			name:       "6-decimal token, $1 per token",
-			price:      big.NewInt(1e18),
-			decimal:    6,
-			wantResult: new(big.Int).Mul(big.NewInt(1e18), big.NewInt(1e12)), // 1e30
-		},
-		{
-			name:       "0-decimal token, $1 per token",
-			price:      big.NewInt(1e18),
-			decimal:    0,
-			wantResult: new(big.Int).Mul(big.NewInt(1e18), big.NewInt(1e18)), // 1e36
-		},
-		{
-			name:       "36-decimal token, $1 per token",
-			price:      big.NewInt(1e18),
-			decimal:    36,
-			wantResult: big.NewInt(1),
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got := calculateUsdPer1e18TokenAmount(tt.price, tt.decimal)
-			assert.Equal(t, tt.wantResult, got)
 		})
 	}
 }
